better test layout
