+++ test-suite/templates/nodejs-npm/one-dep/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "upm-test-js-project",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "express": "^4.18.2"
+  }
+}
